refactored tests
- Moved the tests codebase from test_unparse.py to common.py.
- Eliminated duplicate test_files().
- Moved the remaining test_files() to common.py.
- Fixed inheritance to prevent rerunning same tests many times.

Right now there is a common test case AstunparseCommonTestCase
that defines all tests but leaves implementations of
assertASTEqual() and check_roundtrip() empty.

And, there are two cases: DumpTestCase and UnparseTestCase
that implement assertASTEqual() and check_roundtrip().

The actual tests are the same as in previous implementation.
